requirements.txt
langchain==0.2.8
langchain-community==0.2.7
langchain-core==0.2.19
langchain-text-splitters==0.2.2
uvicorn==0.27.1
fastapi==0.110.0
langchain_aws==0.1.7
sagemaker==2.211.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3
openpyxl==3.1.3
transformers==4.41.2
PyPDF2==3.0.1
faiss-cpu==1.8.0
boto3==1.34.131
# langchain==0.1.12
pydantic==1.10.14
python-multipart==0.0.9
ecdsa==0.19.0
duckduckgo-search==6.1.9
sentence-transformers==2.3.1
apscheduler==3.10.4